refactor(share modal): remove old share modal and create component in rigth folder and use again
