+++ app/pages/multimedia/[id]/[[slug]].vue
@@ -55,11 +55,12 @@
       :request="[`file`]"
     />
 
-    <lazy-common-share-modal
-      v-if="openShare"
+    <lazy-common-modal-base
       v-model:show-dialog="openShare"
-      :title="contentData.title"
-    />
+      title="Share"
+    >
+      <lazy-common-modal-share />
+    </lazy-common-modal-base>
   </v-container>
 </template>
