+++ app/pages/tutorial/[id]/[[slug]].vue
@@ -180,10 +180,12 @@
       v-model:show-dialog="openCrashReport"
       type-crash-report="tutorial"
     />
-    <CommonShareModal
+    <lazy-common-modal-base
       v-model:show-dialog="openShare"
-      :title="contentData.title"
-    />
+      title="Share"
+    >
+      <lazy-common-modal-share />
+    </lazy-common-modal-base>
   </v-container>
 </template>
